requirements.txt
requests
sentence-transformers
scipy
pymupdf
fastapi
faiss-cpu
SQLAlchemy
pydantic
uvicorn[standard]